🐝 prettify scores table
